requirements.txt
flask==2.3.2
python-dotenv==0.19.2
vonage==3.6.0